Commit brainfuck tools
The foremost tool in this collection is the brainfuck debugger.  It was
written to assist with the 'boring flag checker' problem from RaRCTF
2021, but has good potential for general-purpose use.

The compiler and decompiler are much more niche, given brainfuck is not
typically a compiled language.  They are from the same CTF and, although
highly problem-specific, are kept around for posterity.

A hello world program is saved under templates as a quick sanity check
for the tools as well as for reference purposes, should it become useful.

Signed-off-by: Malfurious <m@lfurio.us>