Add Makefile for shellcode templates
The shell*.asm files are considered the default programs and the
expected use-case for utilizing the templates is to edit these files to
implement the desired shellcode.  I figure that literal shellcode makes
the most sense of what to expect by default.

'make all' will assemble and link the shellcode (so it can actually be
directly executed via the output elf files), and feed the disassembly
into shelltool for use elsewhere.

Signed-off-by: Malfurious <m@lfurio.us>