Add readuntil() and readlineuntil() to Comms
Both new functions check the input for a predicate and keep reading
until the predicate is true.
readuntil() will consume input byte by byte and use the entire string
read to check the predicate.  It will then return that entire string.
readlineuntil() consumes input line by line and only uses the last line
to check the predicate.  The line that satisfies the predicate is all
that is returned.

Signed-off-by: dusoleil <howcansocksbereal@gmail.com>