sploit: Add payload builder module
class Payload is a tool for constructing stack-smash payloads and ROP
chains.  Its design is intended to abstract away some of the more
tedious details of crafting a payload.

Payload utilizes mem.Symtbl internally to optionally manage a collection
of named offsets into its own buffer (these are usually in reference to
entities appended to the payload via its main API).  Alternatively, the
API calls to append any entity will return the address of that entity as
well.

Returned (and looked-up) addresses are relative to the beginning of the
payload by default.  However, when the payload is constructed with a
known base address value, these become absolute.  This is useful for
reusing addresses later in the payload body.

class Placeholder is designed to be functionally compatible with
bytearrays and bytestrings.  When constructed, they take the value of
'zero', according to the current arch config.  This facility enables
some API's to detect whether a dummy value was passed as a required
argument when said argument _may_ be unnecessary in niche situations.

Signed-off-by: Malfurious <m@lfurio.us>
Signed-off-by: dusoleil <howcansocksbereal@gmail.com>