sploit: Add startup banner
This just adds a fancy 'SPLOIT' header to the beginning of Sploit's
startup preamble data.  It has the ability to display a few lines of
text beside itself, but most of the things we've planned to put here are
not available yet, so just the operating mode is printed for now.

The SPLOIT text has a colored stripe which, at the moment, also
indicates the operating mode.  This stripe was originally chosen to
balance out the amount of color present in the preamble text, but I've
grown to like it.

Signed-off-by: Malfurious <m@lfurio.us>
Signed-off-by: dusoleil <howcansocksbereal@gmail.com>