sploit: Automatically shutdown outgoing comms after script execution
A new function, Comm.shutdown(), is added.  It will close only the
stdout stream of the communications backend, potentially making the
termination of the target program more fluid.

The name 'shutdown' is chosen to emulate shutdown(2) from the low-level
socket api, which is used to close just part of a full-duplex file
descriptor.  This is in contrast to 'close', which I would expect to
completely terminate the given object IO.

comm.shutdown() is now called by main.py, after the user script returns,
to ensure that the subsequent readall() doesn't get stuck because our
target is blocked reading its stdin.

Signed-off-by: Malfurious <m@lfurio.us>
Signed-off-by: dusoleil <howcansocksbereal@gmail.com>